Initial and final Commit: QueryResults stores information about document and its score.  This is used to rank documents based on query relevance.
QueryStruct stores document's information to process the query.

EasySearch takes in a string to be used as query or "New York" as the
default. The application does not display document relevance, but it
shows the score for the documents and score for each term in the query.

SearchTRECTopics makes use of the previous implementation (TF IDF), to
provide best 1000 documents (most relevant).

CompareAlgorithms performs ranking for 4 different algorithms (Vector
Space, BM25, Dirichlet, and Jelinek). The data is stored in such a way
that can later be evaluated by trec_eval.